added compileVars support and fixed bugs
added __helperTag to helpers so that the parser can auto execute functions when used in syntax like div(hr, p('foo'), hr) where the user did not execute hr. needed identifier to know its a helper and not a normal user func. (we was mistakenly passing compileVars to hr before)

added template.add to replace the uglier template.srcList.push, so that can be extended in future to do other things if needed.